Cleanum of valuations file
