+++ src/storm/storage/umb/model/Valuations.h
@@ -25,6 +25,36 @@
 
 namespace storm::umb {
 
+/*!
+ * Concept for a callback used in readCallback / readValue.
+ * The callback is invoked as callback(entity, variable, value) where value is one of the
+ * supported value types (bool, int64_t, uint64_t, double, RationalNumber, string_view, string,
+ * or std::nullopt_t for unset optional variables).
+ */
+template<typename F>
+concept ValuationReadCallback =
+    std::invocable<F, uint64_t, storm::expressions::Variable const&, bool> || std::invocable<F, uint64_t, storm::expressions::Variable const&, int64_t> ||
+    std::invocable<F, uint64_t, storm::expressions::Variable const&, uint64_t> || std::invocable<F, uint64_t, storm::expressions::Variable const&, double> ||
+    std::invocable<F, uint64_t, storm::expressions::Variable const&, storm::RationalNumber> ||
+    std::invocable<F, uint64_t, storm::expressions::Variable const&, storm::NumberTraits<storm::RationalNumber>::IntegerType> ||
+    std::invocable<F, uint64_t, storm::expressions::Variable const&, std::string_view> ||
+    std::invocable<F, uint64_t, storm::expressions::Variable const&, std::string> ||
+    std::invocable<F, uint64_t, storm::expressions::Variable const&, std::nullopt_t>;
+
+/*!
+ * Concept for a callback used in writeCallback / writeValue.
+ * The callback is invoked as callback(entity, variable, value) where value is passed by
+ * (non-const) lvalue reference so the callback can assign the new value.
+ * For optional variables (AllowOptional=true), value is wrapped in std::optional<ValueType>&.
+ */
+template<typename F>
+concept ValuationWriteCallback =
+    std::invocable<F, uint64_t, storm::expressions::Variable const&, bool&> || std::invocable<F, uint64_t, storm::expressions::Variable const&, int64_t&> ||
+    std::invocable<F, uint64_t, storm::expressions::Variable const&, uint64_t&> || std::invocable<F, uint64_t, storm::expressions::Variable const&, double&> ||
+    std::invocable<F, uint64_t, storm::expressions::Variable const&, storm::RationalNumber&> ||
+    std::invocable<F, uint64_t, storm::expressions::Variable const&, storm::NumberTraits<storm::RationalNumber>::IntegerType&> ||
+    std::invocable<F, uint64_t, storm::expressions::Variable const&, std::string&>;
+
 class Valuations {
    public:
     using Integer = storm::NumberTraits<storm::RationalNumber>::IntegerType;
@@ -34,69 +64,11 @@ class Valuations {
     Valuations& operator=(Valuations const&) = default;
     Valuations& operator=(Valuations&&) = default;
 
-    Valuations(uint64_t const numEntities, std::vector<ValuationClassDescription> const& descriptions, std::vector<char> valuations,
+    Valuations(uint64_t numEntities, std::vector<ValuationClassDescription> const& descriptions, std::vector<char> valuations,
                std::vector<uint64_t> stringMapping, std::vector<char> strings, std::optional<std::vector<uint32_t>> classes = {},
-               std::vector<std::shared_ptr<storm::expressions::ExpressionManager const>> expressionManagers = {})
-        : numEntities(numEntities), valuations(std::move(valuations)), stringMapping(std::move(stringMapping)), strings(std::move(strings)) {
-        STORM_LOG_ASSERT(descriptions.size() == expressionManagers.size() || expressionManagers.size() <= 1,
-                         "Mismatch between number of descriptions and expression managers.");
-        // First set up the variable classes.
-        // We either have a separate manager for each class or all classes share the same manager.
-        // Furthermore, we might create a new manager for a class, if no manager was given explicitly.
-        auto sharedManager = std::make_shared<storm::expressions::ExpressionManager>();
-        for (uint64_t i = 0; i < descriptions.size(); ++i) {
-            if (expressionManagers.empty() || (expressionManagers.size() == 1 && expressionManagers.front() == nullptr)) {
-                // Shared manager for all classes, not given explicitly
-                variableClasses.push_back(createVariablesInformation(*sharedManager, descriptions[i]));
-            } else if (expressionManagers.size() == 1) {
-                // Shared manager for all classes, given explicitly
-                variableClasses.push_back(createVariablesInformation(*expressionManagers.front(), descriptions[i]));
-            } else if (expressionManagers[i] == nullptr) {
-                // Separate manager for each class, not given explicitly
-                auto manager = std::make_shared<storm::expressions::ExpressionManager>();
-                variableClasses.push_back(createVariablesInformation(*manager, descriptions[i]));
-            } else {
-                // Separate manager for each class, given explicitly
-                variableClasses.push_back(createVariablesInformation(*expressionManagers[i], descriptions[i]));
-            }
-        }
-        // Initialize string mappings. It should be given iff there is a string variable
-        bool const hasStringVariable =
-            std::any_of(descriptions.begin(), descriptions.end(), [](auto const& classDescr) { return classDescr.hasStringVariable(); });
-        if (hasStringVariable && this->stringMapping.empty()) {
-            this->stringMapping.push_back(0);
-        }
-        STORM_LOG_ASSERT(hasStringVariable || this->stringMapping.empty(), "Non-empty string mapping given but there is no string variable.");
-        STORM_LOG_ASSERT(stringMapping.empty() || this->stringMapping.back() == strings.size(),
-                         "String mapping should end with the total size of the string data.");
+               std::vector<std::shared_ptr<storm::expressions::ExpressionManager const>> expressionManagers = {});
 
-        // Enable quick access to the right byte span for each entity.
-        if (classes.has_value() && this->variableClasses.size() > 1) {
-            STORM_LOG_ASSERT(numEntities == classes->size(), "Number of entities does not match class mapping size.");
-            this->entityClassMappings = {std::move(*classes), std::vector<uint64_t>({0ull})};
-            uint64_t pos = 0;
-            this->entityClassMappings->toValuationsMapping.reserve(this->entityClassMappings->toClassMapping.size() + 1);
-            for (uint64_t entity = 0; entity < this->entityClassMappings->toClassMapping.size(); ++entity) {
-                STORM_LOG_ASSERT(
-                    this->entityClassMappings->toClassMapping[entity] < this->variableClasses.size(),
-                    "Class index " << this->entityClassMappings->toClassMapping[entity] << " out of bounds. Only " << descriptions.size() << "classes known.");
-                pos += this->variableClasses[this->entityClassMappings->toClassMapping[entity]].sizeInBytes;
-                this->entityClassMappings->toValuationsMapping.push_back(pos);
-            }
-            STORM_LOG_ASSERT(this->valuations.size() == pos, "Valuation data size does not match class mapping.");
-        } else {
-            STORM_LOG_ASSERT(this->variableClasses.size() == 1, "Valuation descriptions must be unique if no class mapping is given.");
-            STORM_LOG_ASSERT(!classes.has_value() || std::all_of(classes->begin(), classes->end(), [&](auto classIndex) { return classIndex == 0; }),
-                             "A single description is given but the class mapping is not unique.");
-            STORM_LOG_ASSERT(this->variableClasses.front().sizeInBytes == 0 || this->valuations.size() % this->variableClasses.front().sizeInBytes == 0,
-                             "Valuation data size is not a multiple of the unique valuation size.");
-            STORM_LOG_ASSERT(numEntities * this->variableClasses.front().sizeInBytes == this->valuations.size(),
-                             "Valuation data size (" << valuations.size() << ") does not match number of entities (" << this->numEntities
-                                                     << ") times valuation size (" << this->variableClasses.front().sizeInBytes << ").");
-        }
-    }
-
-    Valuations(uint64_t const numEntities, ValuationClassDescription const& description, std::vector<char> valuations,
+    Valuations(uint64_t numEntities, ValuationClassDescription const& description, std::vector<char> valuations,
                std::shared_ptr<storm::expressions::ExpressionManager const> expressionManager = {})
         : Valuations(numEntities, {description}, std::move(valuations), {}, {}, std::nullopt, {expressionManager}) {
         STORM_LOG_ASSERT(!description.hasStringVariable(), "String mapping must be given for descriptions with string variables.");
@@ -104,14 +76,10 @@ class Valuations {
 
     Valuations(std::vector<ValuationClassDescription> const& descriptions,
                std::vector<std::shared_ptr<storm::expressions::ExpressionManager const>> expressionManagers = {})
-        : Valuations(0, descriptions, {}, {}, {}, std::vector<uint32_t>{}, expressionManagers) {
-        // Intentionally empty
-    }
+        : Valuations(0, descriptions, {}, {}, {}, std::vector<uint32_t>{}, expressionManagers) {}
 
     Valuations(ValuationClassDescription const& description, std::shared_ptr<storm::expressions::ExpressionManager const> expressionManager = {})
-        : Valuations(0, {description}, {}, {}, {}, std::nullopt, {expressionManager}) {
-        // Intentionally empty
-    }
+        : Valuations(0, {description}, {}, {}, {}, std::nullopt, {expressionManager}) {}
 
     /*!
      * @return the number of entities (e.g. states) that this valuation assigns values for
@@ -124,63 +92,12 @@ class Valuations {
         return variableClasses.size();
     }
 
-    ValuationClassDescription getClassDescription(uint64_t classIndex = 0) const {
-        STORM_LOG_ASSERT(classIndex < numClasses(), "Class index " << classIndex << " out of bounds. Only " << variableClasses.size() << "classes known.");
-        ValuationClassDescription res;
-        uint64_t currBit = 0;
-        for (auto const& varInfo : variableClasses[classIndex].variables) {
-            uint64_t padding = varInfo.bitOffset - currBit;
-            if (varInfo.description.isOptional.value_or(false)) {
-                STORM_LOG_ASSERT(padding >= 1, "Optional variables must have at least 1 bit preceding its offset.");
-                --padding;
-            }
-            if (padding > 0) {
-                res.variables.push_back(storm::umb::ValuationClassDescription::Padding(padding));
-            }
-            res.variables.push_back(varInfo.description);
-            currBit = varInfo.bitOffset + varInfo.description.type.bitSize();
-            STORM_LOG_ASSERT(currBit == res.sizeInBits(), "Unexpected bit offset for variable " << varInfo.description.name << " in class " << classIndex
-                                                                                                << ". Expected " << res.sizeInBits() << ", got "
-                                                                                                << varInfo.bitOffset << ".");
-        }
-        if (uint64_t padding = currBit % 8; padding > 0) {
-            res.variables.push_back(storm::umb::ValuationClassDescription::Padding(8 - padding));
-        }
-        return res;
-    }
+    ValuationClassDescription getClassDescription(uint64_t classIndex = 0) const;
 
-    uint64_t getClassOfEntity(uint64_t entity) const {
-        STORM_LOG_ASSERT(entity < size(), "Entity index out of bounds: " << entity << " >= " << size() << ".");
-        if (entityClassMappings) {
-            return entityClassMappings->toClassMapping[entity];
-        } else {
-            STORM_LOG_ASSERT(variableClasses.size() == 1, "No class mapping given but multiple classes exist.");
-            return 0;
-        }
-    }
+    uint64_t getClassOfEntity(uint64_t entity) const;
 
-    std::span<char const> getRawBytes(uint64_t entity) const {
-        STORM_LOG_ASSERT(entity < size(), "Entity index out of bounds: " << entity << " >= " << size() << ".");
-        if (entityClassMappings) {
-            auto const start = entityClassMappings->toValuationsMapping[entity];
-            auto const end = entityClassMappings->toValuationsMapping[entity + 1];
-            return std::span<char const>(&valuations[start], end - start);
-        } else {
-            auto const start = entity * variableClasses.front().sizeInBytes;
-            return std::span<char const>(&valuations[start], variableClasses.front().sizeInBytes);
-        }
-    }
-
-    std::span<char> getRawBytes(uint64_t entity) {
-        if (entityClassMappings) {
-            auto const start = entityClassMappings->toValuationsMapping[entity];
-            auto const end = entityClassMappings->toValuationsMapping[entity + 1];
-            return std::span<char>(&valuations[start], end - start);
-        } else {
-            auto const start = entity * variableClasses.front().sizeInBytes;
-            return std::span<char>(&valuations[start], variableClasses.front().sizeInBytes);
-        }
-    }
+    std::span<char const> getRawBytes(uint64_t entity) const;
+    std::span<char> getRawBytes(uint64_t entity);
 
     uint64_t numStrings() const {
         return stringMapping.size() > 0 ? stringMapping.size() - 1 : 0;
@@ -190,59 +107,15 @@ class Valuations {
         return !stringMapping.empty();
     }
 
-    void resize(uint64_t newEntityCount, uint64_t const classIndex = 0) {
-        if (newEntityCount > size()) {
-            // Initialize one new entity with default values. This is required to ensure that valuation data is consistent (e.g. avoid 0/0 for rationals).
-            emplaceBack<true>(classIndex, [](auto&&...) {});
-            // For the remaining entities, we can be a bit quicker by copying the values of the last initialized entity.
-            if (newEntityCount > size()) {
-                uint64_t const classSize = variableClasses[classIndex].sizeInBytes;
-                valuations.resize(valuations.size() + (newEntityCount - size()) * classSize);
-                if (entityClassMappings) {
-                    entityClassMappings->toClassMapping.resize(newEntityCount, classIndex);
-                    for (uint64_t valEnd = entityClassMappings->toValuationsMapping.back() + classSize; valEnd < valuations.size(); valEnd += classSize) {
-                        entityClassMappings->toValuationsMapping.push_back(valEnd);
-                    }
-                }
-                auto const srcBytes = getRawBytes(numEntities);  // the bytes of the entry we initialized using emplaceBack
-                for (uint64_t newEntityIndex = numEntities + 1; newEntityIndex < newEntityCount; ++newEntityIndex) {
-                    auto destBytes = getRawBytes(newEntityIndex);
-                    std::copy(srcBytes.begin(), srcBytes.end(), destBytes.begin());
-                }
-                numEntities = newEntityCount;
-            }
-        } else if (newEntityCount < size()) {
-            uint64_t const newValuationsSize = entityClassMappings.has_value() ? entityClassMappings->toValuationsMapping[newEntityCount]
-                                                                               : newEntityCount * variableClasses.front().sizeInBytes;
-            valuations.resize(newValuationsSize);
-            if (entityClassMappings) {
-                entityClassMappings->toClassMapping.resize(newEntityCount);
-                entityClassMappings->toValuationsMapping.resize(newEntityCount + 1);
-            }
-            numEntities = newEntityCount;
-        }
-    }
+    void resize(uint64_t newEntityCount, uint64_t classIndex = 0);
 
-    storm::expressions::ExpressionManager const& getManager() const {
-        auto const& manager = variableClasses.front().expressionManager;
-        STORM_LOG_THROW(
-            std::all_of(variableClasses.begin(), variableClasses.end(), [&manager](auto const& varClass) { return varClass.expressionManager == manager; }),
-            storm::exceptions::IllegalFunctionCallException, "Expression manager is not unique.");
-        return *manager;
-    }
+    storm::expressions::ExpressionManager const& getManager() const;
+    storm::expressions::ExpressionManager const& getManager(uint64_t classIndex) const;
 
-    storm::expressions::ExpressionManager const& getManager(uint64_t classIndex) const {
+    template<bool AllowOptional = false, typename... AllowedTypes, ValuationWriteCallback Callback>
+    void emplaceBack(uint64_t classIndex, Callback const& callback) {
         STORM_LOG_ASSERT(classIndex < variableClasses.size(),
                          "Class index " << classIndex << " out of bounds. Only " << variableClasses.size() << "classes known.");
-        return *variableClasses[classIndex].expressionManager;
-    }
-
-    template<bool AllowOptional = false, typename... AllowedTypes>
-    void emplaceBack(uint64_t const classIndex, auto const& callback) {
-        STORM_LOG_ASSERT(classIndex < variableClasses.size(),
-                         "Class index " << classIndex << " out of bounds. Only " << variableClasses.size() << "classes known.");
-        // Enlarge the valuation data by one entry and initialize it with zeros.
-        // This ensures a consistent state of valuation data, in particular padding bits will always be 0 this way.
         valuations.resize(valuations.size() + variableClasses[classIndex].sizeInBytes, 0);
         if (entityClassMappings) {
             entityClassMappings->toClassMapping.push_back(classIndex);
@@ -254,34 +127,25 @@ class Valuations {
 
    private:
     struct VariableInformation;
-    VariableInformation const& getVariableInformation(uint64_t entity, storm::expressions::Variable const& variable) const {
-        auto const& vars = info(entity).variables;
-        auto varInfoIt = std::find_if(vars.begin(), vars.end(), [&variable](auto const& varInfo) { return varInfo.expressionVariable == variable; });
-        STORM_LOG_ASSERT(varInfoIt != vars.end(), "Can not find unknown variable " << variable.getName() << ".");
-        return *varInfoIt;
-    }
-
-    VariableInformation const& getVariableInformation(storm::expressions::Variable const& variable) const {
-        STORM_LOG_ASSERT(numClasses() == 1, "Trying to get variable information but the class is not unique among entities.");
-        return getVariableInformation(0, variable);
-    }
+    VariableInformation const& getVariableInformation(uint64_t entity, storm::expressions::Variable const& variable) const;
+    VariableInformation const& getVariableInformation(storm::expressions::Variable const& variable) const;
 
    public:
-    template<bool AllowOptional = false, typename... AllowedTypes>
-    void emplaceBack(auto const& callback) {
+    template<bool AllowOptional = false, typename... AllowedTypes, ValuationWriteCallback Callback>
+    void emplaceBack(Callback const& callback) {
         STORM_LOG_ASSERT(variableClasses.size() == 1, "Trying to add a valuation but the class is not unique.");
         emplaceBack<AllowOptional, AllowedTypes...>(0, callback);
     }
 
-    template<typename... AllowedTypes>
-    void readCallback(uint64_t entity, auto const& callback) const {
+    template<typename... AllowedTypes, ValuationReadCallback Callback>
+    void readCallback(uint64_t entity, Callback const& callback) const {
         for (auto const& varInfo : info(entity).variables) {
             read<AllowedTypes...>(entity, varInfo, callback);
         }
     }
 
-    template<typename... AllowedTypes>
-    void readCallback(uint64_t entity, storm::expressions::Variable const& variable, auto const& callback) const {
+    template<typename... AllowedTypes, ValuationReadCallback Callback>
+    void readCallback(uint64_t entity, storm::expressions::Variable const& variable, Callback const& callback) const {
         read<AllowedTypes...>(entity, getVariableInformation(entity, variable), callback);
     }
 
@@ -292,8 +156,8 @@ class Valuations {
         return result;
     }
 
-    template<typename... AllowedTypes>
-    void readCallback(storm::expressions::Variable const& variable, auto const& callback) const {
+    template<typename... AllowedTypes, ValuationReadCallback Callback>
+    void readCallback(storm::expressions::Variable const& variable, Callback const& callback) const {
         if (numClasses() == 1) {
             // We have only one class, so we can look up the variable info once and use it for all entities
             auto const& varInfo = getVariableInformation(variable);
@@ -301,39 +165,44 @@ class Valuations {
                 read<AllowedTypes...>(entity, varInfo, callback);
             }
         } else {
-            // We have multiple classes, so we need to look up the variable info for each entity separately
             for (uint64_t entity = 0; entity < size(); ++entity) {
                 readCallback<AllowedTypes...>(entity, variable, callback);
             }
         }
     }
 
-    template<typename... AllowedTypes>
-    void readCallback(auto const& callback) const {
+    template<typename... AllowedTypes, ValuationReadCallback Callback>
+    void readCallback(Callback const& callback) const {
         for (uint64_t entity = 0; entity < size(); ++entity) {
             readCallback<AllowedTypes...>(entity, callback);
         }
     }
 
-    template<bool InitializeWithCurrent = false, bool AllowOptional = false, typename... AllowedTypes>
-    void writeCallback(uint64_t entity, auto const& callback) {
+    template<bool InitializeWithCurrent = false, bool AllowOptional = false, typename... AllowedTypes, ValuationWriteCallback Callback>
+    void writeCallback(uint64_t entity, Callback const& callback) {
         for (auto const& varInfo : info(entity).variables) {
             write<InitializeWithCurrent, AllowOptional, AllowedTypes...>(entity, varInfo, callback);
         }
     }
 
-    template<bool InitializeWithCurrent = false, bool AllowOptional = false, typename... AllowedTypes>
-    void writeCallback(uint64_t entity, storm::expressions::Variable const& variable, auto const& callback) {
+    template<bool InitializeWithCurrent = false, bool AllowOptional = false, typename... AllowedTypes, ValuationWriteCallback Callback>
+    void writeCallback(uint64_t entity, storm::expressions::Variable const& variable, Callback const& callback) {
         write<InitializeWithCurrent, AllowOptional, AllowedTypes...>(entity, getVariableInformation(entity, variable), callback);
     }
 
     template<typename ValueType>
     void writeValue(uint64_t entity, storm::expressions::Variable const& variable, ValueType const& value) {
-        writeCallback<false, false, ValueType>(entity, variable, [&value](auto const& e, auto const& var, ValueType& val) { val = value; });
+        if constexpr (std::is_same_v<ValueType, std::nullopt_t>) {
+            writeCallback<false, true>(entity, variable, [&value](auto, auto, auto&) { /* intentionally empty */ });
+        } else if constexpr (std::is_same_v<ValueType, std::string_view>) {
+            writeCallback<false, false, std::string>(entity, variable, [&value](auto, auto, std::string& val) { val = value; });
+        } else {
+            writeCallback<false, false, ValueType>(entity, variable, [&value](auto, auto, ValueType& val) { val = value; });
+        }
     }
 
-    template<bool InitializeWithCurrent = false, bool AllowOptional = false, typename... AllowedTypes>
-    void writeCallback(auto const& callback) {
+    template<bool InitializeWithCurrent = false, bool AllowOptional = false, typename... AllowedTypes, ValuationWriteCallback Callback>
+    void writeCallback(Callback const& callback) {
         for (uint64_t entity = 0; entity < size(); ++entity) {
             writeCallback<InitializeWithCurrent, AllowOptional, AllowedTypes...>(entity, callback);
         }
@@ -341,8 +210,6 @@ class Valuations {
 
     /*!
      * Constructs a new Valuations object containing only the selected entities in the given order.
-     * @param selectedEntities
-     * @return
      */
     Valuations selectEntities(auto const& selectedEntities) const {
         Valuations result(variableClasses);
@@ -376,18 +243,7 @@ class Valuations {
         return result;
     }
 
-    typename storm::umb::UmbModel::Valuation getRawUmbData() const {
-        storm::umb::UmbModel::Valuation result;
-        if (entityClassMappings.has_value()) {
-            result.valuationToClass = entityClassMappings->toClassMapping;
-        }
-        result.valuations = valuations;
-        if (hasStrings()) {
-            result.stringMapping = stringMapping;
-            result.strings = strings;
-        }
-        return result;
-    }
+    typename storm::umb::UmbModel::Valuation getRawUmbData() const;
 
    private:
     uint64_t numEntities;
@@ -419,75 +275,13 @@ class Valuations {
     std::vector<uint64_t> stringMapping;
     std::vector<char> strings;
 
-    Valuations(std::vector<VariablesInformation> const& variableClasses) : variableClasses(variableClasses), numEntities(0) {
-        // Intentionally empty
-    }
+    explicit Valuations(std::vector<VariablesInformation> const& variableClasses) : variableClasses(variableClasses), numEntities(0) {}
 
     VariablesInformation const& info(uint64_t entity) const {
         return variableClasses[getClassOfEntity(entity)];
     }
 
-    /*!
-     * Return true if the variable representation and potential offset addition all fit inside a standard 64 bit number representation
-     * @return
-     */
-    static bool fits64Bit(ValuationClassDescription::Variable const& varDesc) {
-        using enum storm::umb::Type;
-        if (varDesc.type.bitSize() > 64) {
-            return false;
-        }
-        switch (varDesc.type.type) {
-            case Bool:
-                return true;
-            case Uint: {
-                // The smallest possible value is 0 + offset. As the offset is given as int64_t, it always fits into 64 bits
-                // We have to check if the largest possible value also fits.
-                if (varDesc.offset.value_or(0) == 0) {
-                    // The offset is 0 and the bitSize is <= 64, so this always fits
-                    return true;
-                } else if (varDesc.upper.has_value()) {
-                    // If the upper bound is given (as int64_t), then the largest value to represent is max(upper, upper - offset).
-                    // Since all numbers involved are given as int64_t, the max value will fit into 64 bits as well.
-                    return true;
-                } else if (varDesc.type.bitSize() == 64) {
-                    // The largest actual value is 2^64 - 1 + offset.
-                    // This never fits into 64 bits for positive offset.
-                    // For negative offsets, the actual value type would be int64_t. Then the above number only fits if offset = -2^63
-                    // Since offset != 0 in this branch, that is the only valid case where the values fit into 64 bits.
-                    return varDesc.offset.value() == std::numeric_limits<int64_t>::min();
-                } else {
-                    // The largest actual value is at most 2^63 - 1 + offset
-                    // For positive offset, the actual type is uint64_t and we can upper bound the above number by 2^63 - 1 + 2^63 - 1 = 2^64 - 2 < uint64_t max
-                    // For negative offset, the actual type is int64_t and we can upper bound the above number by 2^63 - 1 <= int64_t max
-                    return true;
-                }
-            }
-            case Int: {
-                if (varDesc.offset.value_or(0) == 0) {
-                    return true;
-                } else if (varDesc.offset.value() < 0) {
-                    // negative offset. We might have trouble representing the smallest actual value
-                    int64_t const minValueStored =
-                        varDesc.type.bitSize() == 64 ? std::numeric_limits<int64_t>::min() : -(static_cast<int64_t>(1) << (varDesc.type.bitSize() - 1));
-                    return varDesc.lower.has_value() || minValueStored >= std::numeric_limits<int64_t>::min() - varDesc.offset.value();
-                } else {
-                    // positive offset. We might have trouble representing the largest actual value
-                    int64_t const maxValueStored =
-                        varDesc.type.bitSize() == 64 ? std::numeric_limits<int64_t>::max() : (static_cast<int64_t>(1) << (varDesc.type.bitSize() - 1)) - 1;
-                    return varDesc.upper.has_value() || maxValueStored <= std::numeric_limits<int64_t>::max() - varDesc.offset.value();
-                }
-            }
-            case Double:
-                return true;  // double values are always stored as 64 bit IEEE 754 values
-            case Rational:
-                return false;  //
-            case String:
-                return true;  // string indices are always stored as uint64_t
-            default:
-                STORM_LOG_THROW(false, storm::exceptions::NotSupportedException,
-                                "Valuations for variable type '" << varDesc.type.toString() << "' are not supported.");
-        }
-    }
+    static bool fits64Bit(ValuationClassDescription::Variable const& varDesc);
 
     static VariablesInformation createVariablesInformation(auto& expressionManager, ValuationClassDescription const& description) {
         std::vector<VariableInformation> variables;
@@ -539,81 +333,10 @@ class Valuations {
             .variables = std::move(variables), .expressionManager = expressionManager.shared_from_this(), .sizeInBytes = currentOffset / 8};
     }
 
-    bool readBit(std::span<char const> bytes, uint64_t const position) const {
-        STORM_LOG_ASSERT(position < bytes.size() * 8, "Bit position exceeds valuation size.");
-        return bytes[position / 8] & (1 << (position % 8));
-    }
-
-    void writeBit(std::span<char> bytes, uint64_t const position, bool value) const {
-        STORM_LOG_ASSERT(position < bytes.size() * 8, "Bit position exceeds valuation size.");
-        char& byte = bytes[position / 8];
-        char const pos = (1 << (position % 8));
-        if (value) {
-            byte |= pos;
-        } else {
-            byte &= ~pos;
-        }
-    }
-
-    uint64_t readUint64(std::span<char const> bytes, uint64_t const bitOffset, uint64_t const bitSize) const {
-        STORM_LOG_ASSERT(bitOffset < bytes.size() * 8, "Variable offset exceeds valuation size.");
-        STORM_LOG_ASSERT(bitSize <= 64, "Invalid bit range.");
-        auto const firstByte = bitOffset / 8;
-        auto const bitOffsetWithinByte = bitOffset % 8;
-        auto const numBytes = (bitOffsetWithinByte + bitSize + 7) / 8;
-        STORM_LOG_ASSERT(numBytes <= 9, "Invalid number of bytes computed: " << numBytes);
-        uint64_t result;
-        // set the first (up to) 8 bytes
-        std::memcpy(&result, &bytes[firstByte], std::min<uint64_t>(numBytes, 8ull));
-        result >>= bitOffsetWithinByte;
-        // if necessary, set the most significant bits by reading a 9th byte
-        if (numBytes == 9ull) {
-            uint64_t upperBits = std::bit_cast<uint8_t>(bytes[firstByte + 8]);
-            upperBits <<= (64 - bitOffsetWithinByte);
-            result |= upperBits;
-        }
-        // Set irrelevant bits to zero
-        if (bitSize < 64) {
-            uint64_t const relevantBitMask = (1ull << bitSize) - 1;
-            result &= relevantBitMask;
-        }
-        return result;
-    }
-
-    void writeUint64(std::span<char> bytes, uint64_t const bitOffset, uint64_t const bitSize, uint64_t const value) const {
-        STORM_LOG_ASSERT(bitOffset < bytes.size() * 8, "Variable offset exceeds valuation size.");
-        STORM_LOG_ASSERT(bitSize <= 64, "Invalid bit range.");
-        STORM_LOG_ASSERT(bitSize == 64 || value < (1ull << bitSize), "Invalid value " << value << " for bit size " << bitSize);
-        uint64_t const firstByte = bitOffset / 8;
-        uint8_t const bitOffsetWithinByte = bitOffset % 8;
-        uint8_t const numBytes = (bitOffsetWithinByte + bitSize + 7) / 8;
-        uint8_t const numFullBytes = (bitOffsetWithinByte + bitSize) / 8;
-        STORM_LOG_ASSERT(numBytes <= 9, "Invalid number of bytes computed: " << numBytes);
-        if (numFullBytes == 0) {
-            // We only have to write into a single byte
-            char& byte = bytes[firstByte];
-            uint8_t const relevantBitsMask = ((1 << bitSize) - 1) << bitOffsetWithinByte;  // e.g. 0000 1110 for bitOffsetWithinByte=1 and bitSize=3
-            byte &= static_cast<char>(~relevantBitsMask);                                  // set relevant bits to zero
-            byte |= static_cast<char>((value << bitOffsetWithinByte) & relevantBitsMask);  // set relevant bits to the value bits
-        } else {
-            // First write all full bytes
-            if (bitOffsetWithinByte == 0) {
-                // Fast path: variable is byte-aligned, so we can directly write all full bytes without bit shifts
-                std::memcpy(&bytes[firstByte], &value, numFullBytes);
-            } else {
-                uint64_t const shiftedValue = static_cast<uint64_t>(bytes[firstByte]) & ((1ull << bitOffsetWithinByte) - 1) | (value << bitOffsetWithinByte);
-                std::memcpy(&bytes[firstByte], &shiftedValue, numFullBytes);
-            }
-            // Then write the last byte if necessary
-            if (numFullBytes != numBytes) {
-                // we have to write a partial byte at the end, so we need to read the existing byte and only overwrite the relevant bits
-                char& lastByte = bytes[firstByte + numFullBytes];
-                uint8_t const numBitsUsedInLastByte = (bitOffsetWithinByte + bitSize) % 8;
-                lastByte &= static_cast<char>((1 << numBitsUsedInLastByte) - 1);                   // set relevant bits to zero
-                lastByte |= static_cast<char>(value >> (numFullBytes * 8 - bitOffsetWithinByte));  // set relevant bits to the value bits
-            }
-        }
-    }
+    bool readBit(std::span<char const> bytes, uint64_t position) const;
+    void writeBit(std::span<char> bytes, uint64_t position, bool value) const;
+    uint64_t readUint64(std::span<char const> bytes, uint64_t bitOffset, uint64_t bitSize) const;
+    void writeUint64(std::span<char> bytes, uint64_t bitOffset, uint64_t bitSize, uint64_t value) const;
 
     template<bool Signed>
     Integer readInteger(std::span<char const> bytes, uint64_t const bitOffset, uint64_t const bitSize) const {
@@ -634,14 +357,13 @@ class Valuations {
 
     /*!
      * Reads the given variable for the given entity and calls the callback with the read value.
-     * @tparam AllowedTypes either empty (allowing all types) or a list of types that are handled in the callback. The following types are considered:
+     * @tparam AllowedTypes either empty (allowing all types) or a list of types that are handled in the callback.
      * @param entity the entity (state/choice/branch/observation index)
      * @param varInfo The info for the given variable
      * @param callback The callback.
-     * @param defaulOptional if true, optional variables that are not set will be given a default value
      */
-    template<typename... AllowedTypes>
-    void read(uint64_t entity, VariableInformation const& varInfo, auto const& callback) const {
+    template<typename... AllowedTypes, ValuationReadCallback Callback>
+    void read(uint64_t entity, VariableInformation const& varInfo, Callback const& callback) const {
         auto invokeCallback = [&entity, &varInfo, &callback](auto&& value) -> bool {
             using ValueType = std::remove_cvref_t<decltype(value)>;
             bool constexpr IsAllowed = (sizeof...(AllowedTypes) == 0) || std::disjunction_v<std::is_same<ValueType, AllowedTypes>...>;
@@ -840,16 +562,16 @@ class Valuations {
         }
     }
 
-    template<bool InitializeWithCurrent = false, bool AllowOptional = false, typename... AllowedTypes>
-    void write(uint64_t entity, VariableInformation const& varInfo, auto const& callback) {
+    template<bool InitializeWithCurrent = false, bool AllowOptional = false, typename... AllowedTypes, ValuationWriteCallback Callback>
+    void write(uint64_t entity, VariableInformation const& varInfo, Callback const& callback) {
         auto invokeCallback = [this, &entity, &varInfo, &callback]<typename ValueType>() -> bool {
             bool constexpr IsAllowed = (sizeof...(AllowedTypes) == 0) || std::disjunction_v<std::is_same<ValueType, AllowedTypes>...>;
             if constexpr (IsAllowed) {
-                // Initialize with current value (if requested)
                 ValueType value;
                 bool isOptional = varInfo.description.isOptional.value_or(false);
                 bool initializeAsUnsetOptional = false;
                 if constexpr (InitializeWithCurrent) {
+                    // Initialize with current value (if requested)
                     read<std::nullopt_t, ValueType>(entity, varInfo, [&value, &initializeAsUnsetOptional](auto..., auto&& currentValue) {
                         using CurrentVT = std::remove_cvref_t<decltype(currentValue)>;
                         if constexpr (std::is_same_v<CurrentVT, ValueType>) {
@@ -862,7 +584,7 @@ class Valuations {
                 } else {
                     initializeAsUnsetOptional = isOptional;
                     if constexpr (std::is_same_v<ValueType, uint64_t> || std::is_same_v<ValueType, int64_t> || std::is_same_v<ValueType, Integer>) {
-                        // Explicitly initialize integer values to the lower bound, 0, or the upper bound (in that order
+                        // Explicitly initialize integer values to the lower bound, 0, or the upper bound (in that order)
                         if (varInfo.description.lower && (!std::is_same_v<ValueType, uint64_t> || varInfo.description.lower.value() >= 0)) {
                             value = storm::utility::convertNumber<ValueType>(varInfo.description.lower.value());
                         } else if (!varInfo.description.upper || varInfo.description.upper >= 0) {
@@ -876,7 +598,6 @@ class Valuations {
                 bool haveToWriteValue = false;
                 if (isOptional) {
                     if constexpr (AllowOptional) {
-                        // invoke callback with std::optional&
                         std::optional<ValueType> optionalValue = initializeAsUnsetOptional ? std::optional<ValueType>() : value;
                         callback(entity, varInfo.expressionVariable, optionalValue);
                         if (optionalValue.has_value()) {
@@ -932,7 +653,7 @@ class Valuations {
                         return;
                     }
                 }
-                // finally take Integer if the 64 bit types are not allowed or insuficient
+                // finally take Integer if the 64 bit types are not allowed or insufficient
                 if (invokeCallback.template operator()<Integer>()) {
                     return;
                 }
